service-1/requirements.txt
django
djangorestframework
grpcio
grpcio-tools
graphene-django
